engine: make mid-run get_value consistent for conveyor/queue models
After a partial run_to, the #625 resting-curr re-eval re-ran Flows
only, which re-executed every pass-driven flow's AssignConstCurr 0
placeholder into the live chunk: a mid-run get_value of a conveyor
outflow, leak, or queue outflow read 0 instead of the pass-computed
rate, and container values were one step stale relative to the side
tables.

Mirror the step prologue at rest: publish container values from the
real side tables, re-eval Flows, then run the coupled pass on CLONED
side tables so real belts and FIFOs never double-advance. The live
resting chunk now equals the row a resumed run eventually saves --
pinned bit-exactly by segmented-vs-uninterrupted run tests (conveyor,
queue, and coupled fixtures with two rests) plus mid-run invariant
tests across pass-driven flows, containers, arrayed elements, and the
coupled shared flow. A preview error (e.g. a mid-run sample re-latch
exceeding the slat bound) restores the pre-preview chunk and
resurfaces on resume; the preview is presentation-only. Ordinary
models are untouched (guarded on plan-list emptiness), and the
vm_alloc gate now differences equal step-count deltas so the
once-per-run_to preview clone stays visible but per-step scaling
regressions are still caught.

conveyor_compile.rs and vm.rs crossed the per-file line cap with the
new tests, so their test modules move to sibling files via the same
#[path] mount the wasmgen splits use.
